+++ WowPacketParser/SQL/Builders/AreaTriggers.cs
@@ -1,0 +1,39 @@
+﻿using System.Linq;
+using WowPacketParser.Enums;
+using WowPacketParser.Misc;
+using WowPacketParser.Store;
+
+namespace WowPacketParser.SQL.Builders
+{
+    [BuilderClass]
+    public static class AreaTriggers
+    {
+        [BuilderMethod]
+        public static string AreaTriggerTemplateData()
+        {
+            if (Storage.AreaTriggerTemplates.IsEmpty())
+                return string.Empty;
+
+            if (!Settings.SQLOutputFlag.HasAnyFlagBit(SQLOutput.areatrigger_template))
+                return string.Empty;
+
+            var templateDb = SQLDatabase.Get(Storage.AreaTriggerTemplates);
+
+            return SQLUtil.Compare(Settings.SQLOrderByKey ? Storage.AreaTriggerTemplates.OrderBy(x => x.Item1.Id).ToArray() : Storage.AreaTriggerTemplates.ToArray(), templateDb, x => string.Empty);
+        }
+
+        [BuilderMethod]
+        public static string AreaTriggerTemplateVerticesData()
+        {
+            if (Storage.AreaTriggerTemplates.IsEmpty())
+                return string.Empty;
+
+            if (!Settings.SQLOutputFlag.HasAnyFlagBit(SQLOutput.areatrigger_template_polygon_vertices))
+                return string.Empty;
+
+            var templateDb = SQLDatabase.Get(Storage.AreaTriggerTemplatesVertices);
+
+            return SQLUtil.Compare(Settings.SQLOrderByKey ? Storage.AreaTriggerTemplatesVertices.OrderBy(x => x.Item1.AreaTriggerId).ToArray() : Storage.AreaTriggerTemplatesVertices.ToArray(), templateDb, x => string.Empty);
+        }
+    }
+}
