+++ WowPacketParser/SQL/Builders/Miscellaneous.cs
@@ -67,34 +67,6 @@ namespace WowPacketParser.SQL.Builders
             return SQLUtil.Compare(Storage.SniffData, templateDb, x => string.Empty);
         }
 
-        [BuilderMethod]
-        public static string AreaTriggerTemplateData()
-        {
-            if (Storage.AreaTriggerTemplates.IsEmpty())
-                return string.Empty;
-
-            if (!Settings.SQLOutputFlag.HasAnyFlagBit(SQLOutput.areatrigger_template))
-                return string.Empty;
-
-            var templateDb = SQLDatabase.Get(Storage.AreaTriggerTemplates);
-
-            return SQLUtil.Compare(Settings.SQLOrderByKey ? Storage.AreaTriggerTemplates.OrderBy(x => x.Item1.Id).ToArray() : Storage.AreaTriggerTemplates.ToArray(), templateDb, x => string.Empty);
-        }
-
-        [BuilderMethod]
-        public static string AreaTriggerTemplateVerticesData()
-        {
-            if (Storage.AreaTriggerTemplates.IsEmpty())
-                return string.Empty;
-
-            if (!Settings.SQLOutputFlag.HasAnyFlagBit(SQLOutput.areatrigger_template_polygon_vertices))
-                return string.Empty;
-
-            var templateDb = SQLDatabase.Get(Storage.AreaTriggerTemplatesVertices);
-
-            return SQLUtil.Compare(Settings.SQLOrderByKey ? Storage.AreaTriggerTemplatesVertices.OrderBy(x => x.Item1.AreaTriggerId).ToArray() : Storage.AreaTriggerTemplatesVertices.ToArray(), templateDb, x => string.Empty);
-        }
-
         // Non-WDB data but nevertheless data that should be saved to gameobject_template
         /*[BuilderMethod(Gameobjects = true)]
         public static string GameobjectTemplateNonWDB(Dictionary<WowGuid, GameObject> gameobjects)
